### 2018-12-21 cee3c57
update reference in paper.md
