### 2018-12-22 #1138 @PingjunChen
PDF cannot be generated. I have checked reference and image insertion, which should be conformed to format. I wonder if I can ask help on the pdf generation. @pjotrp @arfon
