### 2018-12-22 #1138 @PingjunChen
Thanks @arfon ! Glad you generate pdf for me! I tried using "whedon generate pdf", only "Attempting PDF compilation..." shown above, no pdf link generated. I thought it could be format issue, but could not find any, thus asking for help.
