### 2018-12-28 #1138 @pjotrp
@PingjunChen do you mind regenerating the paper without the image? I have included an image in other papers, take a look http://joss.theoj.org/papers/10.21105/joss.00025.

Also there are some English grammar problems - do you mind having a native speaker check (I am not one!) before we start review. Finally, it took me some time to figure out what the paper actually is about. Can you make the introduction somewhat clearer? Say what the goal is of the software and how it helps its users. Next have a section on state-of-the-art and how your tool differs. I know you cover it somewhat, but make it extremely clear. It will help noobs like me ;)
