### 2018-12-28 8065bec
update paper writing
