### 2018-12-29 f253bfb
revise paper.md
