### 2018-12-29 78d5bfc
remove the image
