### 2018-12-29 63ad65b
remove the image fully
