### 2018-12-29 #1138 @arfon
> @PingjunChen & @Arafatk, we will go into review soon. First I would like to have the pdf generation fixed. We'll await @arfon.
