### 2018-12-29 8738bf9
add image to paper.md
