### 2018-12-29 #1138 @PingjunChen
Hi @pjotrp, I have added the schematic diagram to paper.md and the pdf generation works now! Thanks. One small issue is that the figure caption is overlapped with the footnote. Any suggestions to fix that?
