### 2018-12-29 a0ab0df
add more reference paper
