+++ paper/paper.bib
@@ -18,6 +18,25 @@
   organization={International Society for Optics and Photonics}
 }
 
+@inproceedings{cruz2014automatic,
+  title={Automatic detection of invasive ductal carcinoma in whole slide images with convolutional neural networks},
+  author={Cruz-Roa, Angel and Basavanhally, Ajay and Gonz{\'a}lez, Fabio and Gilmore, Hannah and Feldman, Michael and Ganesan, Shridar and Shih, Natalie and Tomaszewski, John and Madabhushi, Anant},
+  booktitle={Medical Imaging 2014: Digital Pathology},
+  volume={9041},
+  pages={904103},
+  year={2014},
+  organization={International Society for Optics and Photonics}
+}
+
+@inproceedings{hou2016patch,
+  title={Patch-based convolutional neural network for whole slide tissue image classification},
+  author={Hou, Le and Samaras, Dimitris and Kurc, Tahsin M and Gao, Yi and Davis, James E and Saltz, Joel H},
+  booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition},
+  pages={2424--2433},
+  year={2016}
+}
+
+
 @inproceedings{chen2018automatic,
   title={Automatic Pathology Diagnosis on Large Slide Image Using Patch Aggregation},
   author={Chen, Pingjun and Xie, Yuanpu and Hai, Su and Yang, Lin},
