### 2018-12-29 14dcff0
redrawn the schematic diagram
