### 2018-12-30 #1138 @pjotrp
The figure caption outside the page is probably a LaTeX problem. Making the figure smaller will fix it. Or have the caption as text so it goes on the next page... @arfon, we probably don't have a way to constrain the figure in a vertical box?
