### 2018-12-30 #1148 @whedon
[REVIEW]: tissueloc: Whole slide pathology image tissue localization
**Submitting author:** @PingjunChen (<a href="http://orcid.org/0000-0003-0528-1713">Pingjun Chen</a>)
**Repository:** <a href="https://github.com/PingjunChen/tissueloc" target ="_blank">https://github.com/PingjunChen/tissueloc</a>
**Version:** v1.0.0
