### 2018-12-30 #1148 @PingjunChen
Hi @pjotrp and @Arafatk, all listed boxes can be ticked except "Automated tests". Because of the gigabyte size of whole slide image (WSI), we think uploading testing WSIs to github repo is not good practice. But tests are conducted extensively in local repo.
