### 2018-12-31 #1148 @pjotrp
@PingjunChen I agree that storing the data in github is not a good idea. What I do is provide data that is readily available for download in some way. A web server or figshare or something would do. But, to be honest, since this is a tool using standard libraries - so I am OK to skip it. If you decide to develop it further I would certainly recommend adding tests. You could also consider including some smaller images - just to make sure it won't regress.
