### 2018-12-31 f2bd32e
add autotest
