### 2018-12-31 1195dd9
update README and travis
