### 2018-12-31 #1148 @PingjunChen
Thanks for your suggestions! @pjotrp I have add a few low level slide images, which is much smaller, for testing tissue localization. They are located in "$project_dir/test/data". With these testing images, now automated test using travis is added and passed.
