### 2018-12-31 6ff53ab
add Figshare for whole slide image testing
