### 2018-12-31 #1148 @Arafatk
@PingjunChen
1.  The installation worked fine but I had to install python-tk separately to run the software.  For a sanity check for installation generally I go to Codeanywhere.io and spawn small virtual machines to try out my installation and see if it works on a fresh instances.
2. I see this on the generated PDF
![image](https://user-images.githubusercontent.com/11155207/50567851-d2c27100-0d6f-11e9-9721-218e3e2ff4e4.png)
Can you please resize the image size to make this look cleaner?
3. Can you please elaborate a little bit on the community guidelines?
Maybe use some other software as a reference point -> https://github.com/SciRuby/scirub/blob/master/CONTRIBUTING.md
4. I think paper.md must also have references just like the generated pdf but maybe I am wrong. -> https://github.com/PingjunChen/tissueloc/blob/master/paper/paper.md
