### 2018-12-31 5381c65
1. remove matplotlib in repo. 2. resize schemic diagram to be smaller. 3. Elaborate community guidelines.
