### 2018-12-31 4999a79
update version to 1.1.0
