### 2018-12-31 34b1b81
update version to 1.2.0
