+++ requirements.txt
@@ -1,4 +1,5 @@
+numpy==1.15.4
 Pillow==5.3.0
-opencv-python==3.4.4.19
-scikit-image==0.14.1
 openslide-python==1.1.1
+scikit-image==0.14.1
+opencv-python==3.4.4.19
