+++ .travis.yml
@@ -1,14 +1,13 @@
 language: python
 python:
-    - "3.6"
+    - "3.7"
 
 before_install:
     - sudo apt-add-repository -y ppa:ubuntugis/ubuntugis-unstable
     - sudo apt-get update
     - sudo apt-get install openslide-tools
     - pip install --upgrade pip
-    - pip install opencv-python
-    - pip install shapely
+    - pip install requirements.txt
     - pip install -U pytest
 
 script:
