### 2019-01-01 b24827f
travis donot support Python3.7, change to 3.6
