### 2019-01-01 a653d61
update travis to use pip install requirements.txt
