### 2019-01-01 f343faf
update version to v1.3.0
