### 2019-01-01 #1148 @PingjunChen
@Arafatk @pjotrp  By the way, I now upgrade the ``tissueloc`` version to 1.4.0.
