### 2019-01-01 #1148 @pjotrp
Thank you @Arafatk! @PingjunChen can you also discuss the boxes that have not been checked on the check list? Mostly these are minor additions to the README file. They are important for people who want to try the software. Almost there!
