### 2019-01-01 #1148 @PingjunChen
1. For the installation, I have updated the instructions in README.md and remove "import matplotlib" in the repo. I have tried the installation on both my Ubuntu and Mac machine.
2. I use the "travis" and "pytest" for automated test. The status is "passing".
3. I have updated the "Contributing" part. I would like to refer the reference point @Arafatk suggested. Sadly, the link cannot be opened.

Thanks @pjotrp and @Arafatk !
