### 2019-01-01 #1148 @Arafatk
@PingjunChen Thanks for the work. All the suggestions that I made have been taken care off.

By the way, this is the link that I had suggested ->  https://github.com/SciRuby/sciruby/blob/master/CONTRIBUTING.md

Also, this is a very trivial issues but perhaps Contribution part could be phrased a bit differently just for clarity. So

![image](https://user-images.githubusercontent.com/11155207/50572302-0e4a5300-0de3-11e9-953a-353dc54c5473.png)

Might also be written as

We would like to test this package on more diversified digital slides. Slides (low level images would be better) and their corresponding results are also very welcome as [Pull Requests](https://github.com/PingjunChen/tissueloc/pulls).
