### 2019-01-01 38a1d72
update paper.md and resize the attached image
