### 2019-01-01 #1148 @PingjunChen
@Arafatk Thanks for your feedback. I revised the last paragraph in "Contributing" as you suggested. Also Thanks for the link you provide.
