### 2019-01-01 da75465
update schematic diagram caption in paper.md
