### 2019-01-02 #1148 @pjotrp
Thank you @Arafatk and @PingjunChen. Great job.

The review process is now complete. To finalize your submission and accept your paper in JOSS, we need two things. First, can you confirm that all references in your bibliography have a DOI (if one exists).

Second, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here with the DOI for the release.
