### 2019-01-02 685640c
check ref doi and remove extra images
