+++ paper/paper.bib
@@ -35,14 +35,3 @@
   pages={2424--2433},
   year={2016}
 }
-
-
-@inproceedings{chen2018automatic,
-  title={Automatic Pathology Diagnosis on Large Slide Image Using Patch Aggregation},
-  author={Chen, Pingjun and Xie, Yuanpu and Hai, Su and Yang, Lin},
-  booktitle={MODERN PATHOLOGY},
-  volume={31},
-  pages={586--586},
-  year={2018},
-  organization={NATURE PUBLISHING GROUP 75 VARICK ST, 9TH FLR, NEW YORK, NY 10013-1917 USA}
-}
