### 2019-01-02 #1148 @PingjunChen
0. I just reviewed the repo and tagged the repo v1.5.0 as the final version for JOSS.
1. After checking, now I confirm all references in the bibliography have DOI.
2. DOI for the release is 10.5281/zenodo.2529982, with badge shown as
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.2529982.svg)](https://doi.org/10.5281/zenodo.2529982)

Thanks! @pjotrp and @Arafatk !
