### 2019-01-02 #1148 @pjotrp
You need to add the doi for each reference as in https://github.com/genenetwork/genenetwork2/blob/testing/doc/joss/2016/paper.bib so it shows in the paper.
