+++ paper/paper.bib
@@ -5,6 +5,7 @@
   volume={30},
   pages={60--71},
   year={2016},
+  doi={10.1016/j.media.2015.12.002},
   publisher={Elsevier}
 }
 
@@ -15,6 +16,7 @@
   volume={9041},
   pages={904103},
   year={2014},
+  doi={10.1117/12.2043872},
   organization={International Society for Optics and Photonics}
 }
 
@@ -33,16 +35,6 @@
   author={Hou, Le and Samaras, Dimitris and Kurc, Tahsin M and Gao, Yi and Davis, James E and Saltz, Joel H},
   booktitle={Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition},
   pages={2424--2433},
+  doi={10.1109/CVPR.2016.266},
   year={2016}
-}
-
-
-@inproceedings{chen2018automatic,
-  title={Automatic Pathology Diagnosis on Large Slide Image Using Patch Aggregation},
-  author={Chen, Pingjun and Xie, Yuanpu and Hai, Su and Yang, Lin},
-  booktitle={MODERN PATHOLOGY},
-  volume={31},
-  pages={586--586},
-  year={2018},
-  organization={NATURE PUBLISHING GROUP 75 VARICK ST, 9TH FLR, NEW YORK, NY 10013-1917 USA}
 }
