### 2019-01-02 46447f4
add doi to all references
