+++ paper/paper.bib
@@ -27,6 +27,7 @@
   volume={9041},
   pages={904103},
   year={2014},
+  do={10.1117/12.2043872},
   organization={International Society for Optics and Photonics}
 }
 
