+++ paper/paper.bib
@@ -20,16 +20,6 @@
   organization={International Society for Optics and Photonics}
 }
 
-@inproceedings{cruz2014automatic,
-  title={Automatic detection of invasive ductal carcinoma in whole slide images with convolutional neural networks},
-  author={Cruz-Roa, Angel and Basavanhally, Ajay and Gonz{\'a}lez, Fabio and Gilmore, Hannah and Feldman, Michael and Ganesan, Shridar and Shih, Natalie and Tomaszewski, John and Madabhushi, Anant},
-  booktitle={Medical Imaging 2014: Digital Pathology},
-  volume={9041},
-  pages={904103},
-  year={2014},
-  do={10.1117/12.2043872},
-  organization={International Society for Optics and Photonics}
-}
 
 @inproceedings{hou2016patch,
   title={Patch-based convolutional neural network for whole slide tissue image classification},
