### 2019-01-02 #1148 @PingjunChen
@pjotrp
I misunderstood what you mean. Now DOIs are added to each referenced paper.
By the way, shall I release a new version to Zenodo again because of this update?
