### 2019-01-02 #1148 @arfon
@PingjunChen - your paper is now accepted into JOSS :zap: :rocket: :boom:
